refactor(niri): give each fact one home, and swap a frame whole
The module root had kept everything the submodules were carved out of: the
pattern integer helpers, the short-circuit and discriminator rewrites, the
lattice joins. Each moves to the module that asks the question it answers, and
the root drops from 1193 lines to 746.

`AliasInfo` leaves for `optimize::alias`, its only reader, and `LocalSet` for
the arena that defines what a local index is. `LocalSet::is_empty` had no
caller.

The projections in `frame.rs` that take `&self` are reads, which is the other
module's subject; they move to `lattice.rs` so the frame is what performs.

The in-place reduction walked a list of node kinds it maintained itself, and
had fallen behind the IR — a call's arguments, an index, an array literal and
a pattern's constant were all left unwalked. It now takes its children from
`Body::for_each_child`, so a kind added to the IR is walked without anyone
remembering to, and only the two positions that name storage rather than a
value stay special.

A compile-time call had saved and restored five interpreter fields by hand,
with the whole run in between. They become one `FrameState` exchanged in a
single statement, so the caller's frame goes back whatever the run decides and
no `?` can be introduced between the two.

The `CalleeMap` doc named `is_ctfe_eligible` as its safety gate; the only
builder gates on `is_ctfe_runnable`. A frame runs a unit callee for the writes
it performs, so requiring a value would refuse work the frame does — the doc
was describing an older map.
